+++ docs/change_log.rst
@@ -50,6 +50,10 @@ The following are only available within Stem's `git repository
 
   * Socket based control connections often raised BrokenPipeError when closed
 
+ * **Descriptors**
+
+  * *transport* lines within extrainfo descriptors failed to validate
+
 .. _version_1.8:
 
 Version 1.8 (December 29th, 2019)